// types.ts
export interface LogicStatement {
  text: string;
  type: string; // e.g., "Meimra", "Kushya"
  color: string;
  analysis: {
    subject: string;
    predicate: string;
    statementType: string;
    reason?: string;
  }
}

export interface Syllogism {
  premise1: string;
  premise2: string;
  conclusion: string;
}

export interface LogicSystemData {
  statements: LogicStatement[];
  syllogism: Syllogism;
}